Improvements to web scraping capabilities
it now stores all scraped data in a dict which probably will need to be changed. Using class to identify elements but feels sketchy as they look random but is probably fine I hope. Still need to add more info but they do not have as easily accessed classes and i need a bit of a break.